% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbvnorm.R
\name{rbvnorm}
\alias{rbvnorm}
\title{rbvnorm - Gibbs Sampler}
\usage{
rbvnorm(iter, mu, sigma)
}
\arguments{
\item{iter}{number of iterations for gibbs sampler}

\item{mu}{vector of means}

\item{sigma}{covariance matrix}
}
\value{
a list of the gibbs samples, iterations, mu, and sigma
}
\description{
rbvnorm - Gibbs Sampler
}
\examples{
rbvnorm(iter = 10000, mu = c(3,6), sigma = matrix(c(8,-2,-2,4), nrow =2, byrow = TRUE))
}
