% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbsSampler.R
\name{rbvnorm}
\alias{rbvnorm}
\title{Rbvnorm}
\usage{
rbvnorm(iter, mu, sigma, seed = 3)
}
\arguments{
\item{iter}{the number of iterations to make}

\item{mu}{the mean vector of the BVN distribution}

\item{sigma}{the covariance matrix of the BVN distribution}
}
\value{
a named list of a dataframe of x1 and x2 iterations and original parameters
}
\description{
Rbvnorm
}
\examples{
rbvnorm(10, c(1, 1, 1), matrix(c(1, 1, 1, 1, 1, 1, 1, 1, 1),nrow=3),seed=3)
rbvnorm(iter = 10000, mu = c(3,6), sigma = matrix(c(8,-2,-2,4), nrow =2, byrow = TRUE))
}
